+++ bootique-kafka-streams/src/test/java/io/bootique/kafka/streams/KafkaStreamsModuleIT.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to ObjectStyle LLC under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ObjectStyle LLC licenses
+ * this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package io.bootique.kafka.streams;
+
+import io.bootique.BQRuntime;
+import io.bootique.test.junit.BQTestFactory;
+import org.junit.Rule;
+import org.junit.Test;
+
+import static org.junit.Assert.assertNotNull;
+
+public class KafkaStreamsModuleIT {
+
+    @Rule
+    public final BQTestFactory testFactory = new BQTestFactory();
+
+    @Test
+    public void testKafkaStreamsFactory() {
+        BQRuntime runtime = testFactory.app().autoLoadModules().createRuntime();
+        KafkaStreamsFactory factory = runtime.getInstance(KafkaStreamsFactory.class);
+        assertNotNull(factory);
+    }
+}