+++ bootique-kafka-streams/src/test/java/io/bootique/kafka/streams/KafkaStreamsModuleProviderTest.java
@@ -1,0 +1,12 @@
+package io.bootique.kafka.streams;
+
+import io.bootique.test.junit.BQModuleProviderChecker;
+import org.junit.Test;
+
+public class KafkaStreamsModuleProviderTest {
+
+    @Test
+    public void testAutoLoadable() {
+        BQModuleProviderChecker.testAutoLoadable(KafkaStreamsModuleProvider.class);
+    }
+}